requirements.txt
fastapi
uvicorn
langchain
langchain_openai
langchain_community
python-dotenv
faiss-cpu